worker_scripts/requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
